fix: Extract split-run popup data hook for ESLint budget
Move live artifact and pull request loading out of
WorkOrderSplitRunPopup so complexity stays within the CI
lint budget after the pull request UI changes.
