+++ web_src/src/pages/factories/pages/work-order-split-run/WorkOrderSplitRunPopup.tsx
@@ -1,4 +1,3 @@
-import { useFactoryPullRequests, useWorkOrderArtifacts } from "@/hooks/useFactoryData";
 import { useEffect, useMemo, useState } from "react";
 
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
@@ -18,18 +17,13 @@ import {
   type SplitRunPhaseId,
 } from "./splitRunMocks";
 import {
-  collectSplitRunArtifacts,
-  collectSplitRunPullRequests,
   defaultSplitRunPopupTab,
-  resolveSplitRunPopupArtifacts,
-  resolveSplitRunPopupPullRequests,
   type SplitRunPopupTab,
-  splitRunDescriptionMarkdown,
   splitRunLogTabDotClass,
   splitRunPhaseAutomationHref,
   splitRunPhaseRunHref,
-  splitRunSourceDescription,
 } from "./splitRunPopupModel";
+import { useSplitRunPopupData } from "./useSplitRunPopupData";
 import { useSplitRunFooterActions, type SplitRunFooterActions } from "./useSplitRunFooterActions";
 import { useSplitRunWorkOrderEdits } from "./useSplitRunWorkOrderEdits";
 import { useSplitRunLiveCanvas } from "./useSplitRunLiveCanvas";
@@ -210,38 +204,14 @@ export function WorkOrderSplitRunPopup({
 }) {
   const footerActions = useSplitRunFooterActions(organizationId, factoryId, orderId);
   const mutations = footerMutationHandlers(canUpdate, footerActions, fixture);
-  const fixtureArtifacts = collectSplitRunArtifacts(fixture);
-  const fixturePullRequests = collectSplitRunPullRequests(fixture);
-  const useLiveArtifacts = hasLiveWorkOrder(organizationId, factoryId, orderId);
-  const liveArtifactsQuery = useWorkOrderArtifacts(organizationId ?? "", factoryId ?? "", orderId ?? "");
-  const livePullRequestsQuery = useFactoryPullRequests(
-    organizationId ?? "",
-    factoryId ?? "",
-    orderId ? { workOrderIds: [orderId] } : undefined,
-  );
-  const artifacts = resolveSplitRunPopupArtifacts({
-    fixtureArtifacts,
-    liveArtifacts: liveArtifactsQuery.data,
-    useLive: useLiveArtifacts,
-  });
-  const pullRequests = resolveSplitRunPopupPullRequests({
-    fixturePullRequests,
-    livePullRequests: livePullRequestsQuery.data,
-    useLive: useLiveArtifacts,
-  });
-  const artifactDescription = splitRunDescriptionMarkdown(artifacts) || splitRunDescriptionMarkdown(fixtureArtifacts);
-  const sourceDescription = splitRunSourceDescription({
-    workOrderDescription: fixture.descriptionText,
-    artifactDescription,
-    preferWorkOrder: useLiveArtifacts,
-  });
+  const popupData = useSplitRunPopupData({ organizationId, factoryId, orderId, fixture });
   const edits = useSplitRunWorkOrderEdits({
     organizationId,
     factoryId,
     orderId,
     canUpdate,
     title: fixture.title,
-    description: sourceDescription,
+    description: popupData.sourceDescription,
     owner: fixture.owner,
     assigneeIds: fixture.assigneeIds ?? [],
     footerKind: fixture.footer.kind,
@@ -268,11 +238,11 @@ export function WorkOrderSplitRunPopup({
       <SplitRunPopupTabs
         fixture={fixture}
         edits={edits}
-        artifacts={artifacts}
-        artifactsLoading={useLiveArtifacts && liveArtifactsQuery.isLoading}
-        pullRequests={pullRequests}
-        pullRequestsLoading={useLiveArtifacts && livePullRequestsQuery.isLoading}
-        pullRequestsError={useLiveArtifacts ? (livePullRequestsQuery.error ?? null) : null}
+        artifacts={popupData.artifacts}
+        artifactsLoading={popupData.artifactsLoading}
+        pullRequests={popupData.pullRequests}
+        pullRequestsLoading={popupData.pullRequestsLoading}
+        pullRequestsError={popupData.pullRequestsError}
         organizationId={organizationId}
         factoryId={factoryId}
         factoryKey={factoryKey}
@@ -332,10 +302,6 @@ function returnToBacklogAction(
   };
 }
 
-function hasLiveWorkOrder(organizationId?: string, factoryId?: string, orderId?: string) {
-  return Boolean(organizationId && factoryId && orderId);
-}
-
 function SplitRunPopupTabs({
   fixture,
   edits,
@@ -357,9 +323,9 @@ function SplitRunPopupTabs({
 }: {
   fixture: SplitRunFixture;
   edits: ReturnType<typeof useSplitRunWorkOrderEdits>;
-  artifacts: ReturnType<typeof collectSplitRunArtifacts>;
+  artifacts: ReturnType<typeof useSplitRunPopupData>["artifacts"];
   artifactsLoading: boolean;
-  pullRequests: ReturnType<typeof collectSplitRunPullRequests>;
+  pullRequests: ReturnType<typeof useSplitRunPopupData>["pullRequests"];
   pullRequestsLoading: boolean;
   pullRequestsError: Error | null;
   organizationId?: string;
